Create Use of lambda() with map()
Use of lambda() with map()
The map() function in Python returns a list of the results after applying the given function to each item of a given iterable (list, tuple etc.). The general syntax of map() function is map(fun, iter). We can also use lambda functions with map(). Refer to slide 5 of video 1.7 for general help of map() function with lambda().

In this exercise, you'll be using lambda function inside the map() built-in function to square all numbers in the list.

Instructions
Print my_list which is available in your environment.
Square each item in my_list using map() and lambda().
Print the result of map function.