Create Use of lambda() with filter()
Use of lambda() with filter()
Another function that is used extensively in Python is the filter() function. The filter() function in Python takes in a function and a list as arguments. Similar to the map(), filter() can be used with lambda function. Refer to slide 6 of video 1.7 for general help of the filter() function with lambda().

In this exercise, you'll be using lambda() function inside the filter() built-in function to find all the numbers divisible by 10 in the list.

Instructions

Print my_list2 which is available in your environment.
Filter the numbers divisible by 10 from my_list2 using filter() and lambda().
Print the numbers divisible by 10 from my_list2.