Create CountingBykeys
CountingBykeys
For many datasets, it is important to count the number of keys in a key/value dataset. For example, counting the number of countries where the product was sold or to show the most popular baby names. In this simple exercise, you'll use the Rdd that you created earlier and count the number of unique keys in that pair RDD.

Remember, you already have a SparkContext sc and Rdd available in your workspace.

Instructions

countByKeyand assign the result to a variable total.
What is the type of total?
Iterate over the total and print the keys and their counts.